Swappable DFAs with wrapper functions
This will let me make a WTF-8 DFA and use the same logic core for
that, and again (with slightly more tweaking) for no-codepoint
validation.

Also noticed that the count-codepoint code was at risk of reading past
the buffer if a codepoint was truncated, replaced the guts with an
unroll for a modest speedup.